MBR IPL will start the loader from the boot drive
Instead of using a fixed disk number, set at install time, as the disk
from which the XOSL loader is booted, the MBR IPL will now use the boot
drive, as received from the BIOS.

This is intended to handle the following scenario:
* Drive 0x80: USB bootable key with XOSL installer
* Drive 0x81: Disk holding the DOS partition to which XOSL is installed

Normally, XOSL would configure the IPL to look for the loader on drive
0x81, but when the USB key is then removed, disk 0x81 becomes 0x80. The
IPL is then unable to load anything.

By using th boot drive number from the BIOS, the IPL will adjust to the
changing numbering so that thing don't break.

Note that some tweaking to the IPL code was needed to spare some bytes
and keep the whole thing below 440 bytes in size.