Install the MBR IPL to same disk as the partition
Normally XOSL would install the MBR IPL to disk 0x80. This breaks things
when such drive maps to i.e. a recovery USB key with XOSL installer,
which is going to be removed after installation: neither XOSL nor the
key will now boot since XOSL's IPL is now on the USB stick :(

The IPL should really be installed to the same disk that contains the
selected installation partition. This makes the installation independent
of how the BIOS maps drives to numbers.